fix compiler warning
git-svn-id: http://vtp.googlecode.com/svn/trunk@6599 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainSDK/vtdata/MathTypes.cpp
@@ -675,7 +675,7 @@ void  DPolygon2::InsertPointAfter(int iInsertAfter, const DPoint2 &Point)
 	for (unsigned int ring = 0; ring < size(); ring++)
 	{
 		unsigned int size = at(ring).GetSize();
-		if (iInsertAfter < size)
+		if ((unsigned int)iInsertAfter < size)
 		{
 			// remove the point from this ring
 			at(ring).InsertPointAfter(iInsertAfter, Point);
@@ -693,7 +693,7 @@ void DPolygon2::RemovePoint(int N)
 	for (unsigned int ring = 0; ring < size(); ring++)
 	{
 		unsigned int size = at(ring).GetSize();
-		if (N < size)
+		if ((unsigned int)N < size)
 		{
 			// remove the point from this ring
 			at(ring).RemovePoint(N);
